In xlog_redo(), believe a SHUTDOWN checkPoint.oldestMulti exactly.
It was so before this branch.  This restores consistency with the
handling of nextXid, nextMulti and oldestMulti: we treat them as exact
for XLOG_CHECKPOINT_SHUTDOWN and as minima for XLOG_CHECKPOINT_ONLINE.
I do not know of a case where this definitely matters for any of these
counters.  It might matter if a bug causes oldestXid to move forward
wrongly, causing it to then move backward later.  (I don't know if
VACUUM does ever move oldestXid backward, but it's a plausible thing to
do if on-disk state fully agrees with an older value.)  That example has
no counterpart for oldestMultiXactId, because any update first arrives
in an XLOG_MULTIXACT_TRUNCATE_ID record.  Therefore, this commit is
probably cosmetic.